Consolidated functgroups.records with _daylight_scrape
